requirements.txt
eel==0.15.0
pillow==9.5.0